requirements.txt
discord.py
asyncio
datetime
python_dotenv
motor